ai_models/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.7.2
numpy==1.26.4
pandas==2.2.0
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
